Feat: adds functioning heatmap with real IP data
